starting over on character rendering process
Rather than trying to do the full pipeline of character
row buffer/font data/palette data to compute colors of
pixels to add to the pixel buffer, just try to generate
fixed patterns.

Initial progress: it KIND of works for generating a fixed
pattern of 10000000 (1 foreground followed by 7 background
pixels), except that everything sems to be offset one pixel
column to the right.